Add code for spawning bullets
The GameEventListener class now implements a method to update its
current time to be able to check if a bullet can be spawned. Touch
events now provide the ability for another method in the class to
determine the direction of the bullet.

Bullet spawn interval can be adjusted to increase/decrease the frequency
of bullets spawning.